Add support of the ports filtering to the dashboard.

+++ src/main/java/com/exalttech/trex/ui/models/json/stats/streams/JSONFlowStatsStream.java
@@ -4,10 +4,10 @@ import java.util.HashMap;
 
 
 public class JSONFlowStatsStream {
-    HashMap<Integer,Integer> tx_pkts;
-    HashMap<Integer,Integer> tx_bytes;
-    HashMap<Integer,Integer> rx_pkts;
-    HashMap<Integer,Integer> rx_bytes;
+    HashMap<Integer, Long> tx_pkts;
+    HashMap<Integer, Long> tx_bytes;
+    HashMap<Integer, Long> rx_pkts;
+    HashMap<Integer, Long> rx_bytes;
 
     public JSONFlowStatsStream() {
         tx_pkts = new HashMap<>();
@@ -17,10 +17,10 @@ public class JSONFlowStatsStream {
     }
 
     public JSONFlowStatsStream(
-            HashMap<Integer,Integer> tx_pkts,
-            HashMap<Integer,Integer> tx_bytes,
-            HashMap<Integer,Integer> rx_pkts,
-            HashMap<Integer,Integer> rx_bytes
+            HashMap<Integer, Long> tx_pkts,
+            HashMap<Integer, Long> tx_bytes,
+            HashMap<Integer, Long> rx_pkts,
+            HashMap<Integer, Long> rx_bytes
     ) {
         this.tx_pkts = tx_pkts;
         this.tx_bytes = tx_bytes;
@@ -28,35 +28,35 @@ public class JSONFlowStatsStream {
         this.rx_bytes = rx_bytes;
     }
 
-    public HashMap<Integer, Integer> getTx_pkts() {
+    public HashMap<Integer, Long> getTx_pkts() {
         return tx_pkts;
     }
 
-    public void setTx_pkts(HashMap<Integer, Integer> tx_pkts) {
+    public void setTx_pkts(HashMap<Integer, Long> tx_pkts) {
         this.tx_pkts = tx_pkts;
     }
 
-    public HashMap<Integer, Integer> getTx_bytes() {
+    public HashMap<Integer, Long> getTx_bytes() {
         return tx_bytes;
     }
 
-    public void setTx_bytes(HashMap<Integer, Integer> tx_bytes) {
+    public void setTx_bytes(HashMap<Integer, Long> tx_bytes) {
         this.tx_bytes = tx_bytes;
     }
 
-    public HashMap<Integer, Integer> getRx_pkts() {
+    public HashMap<Integer, Long> getRx_pkts() {
         return rx_pkts;
     }
 
-    public void setRx_pkts(HashMap<Integer, Integer> rx_pkts) {
+    public void setRx_pkts(HashMap<Integer, Long> rx_pkts) {
         this.rx_pkts = rx_pkts;
     }
 
-    public HashMap<Integer, Integer> getRx_bytes() {
+    public HashMap<Integer, Long> getRx_bytes() {
         return rx_bytes;
     }
 
-    public void setRx_bytes(HashMap<Integer, Integer> rx_bytes) {
+    public void setRx_bytes(HashMap<Integer, Long> rx_bytes) {
         this.rx_bytes = rx_bytes;
     }
 }
